♻️ Refactor for v1.4.0 🚚 Merge include & src folder 🚚 Merge res and qml folder 🔨 Introduce intermediate library QaterialComponents 📄 Update license in some files
