+++ CMakeLists.txt
@@ -1,6 +1,24 @@
+# MIT License
 #
-#   Qaterial CMake script by Olivier Le Doeuff
+# Copyright (c) 2020 Olivier Le Doeuff
 #
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
 
 cmake_minimum_required(VERSION 3.14.0 FATAL_ERROR)
 
@@ -70,7 +88,6 @@ find_package(Qt5 REQUIRED COMPONENTS
   Xml
   QuickCompiler
 )
-find_package(Qt5 QUIET COMPONENTS QmlWorkerScript)
 
 # ───── FETCH DEPEDENCIES ─────
 
@@ -81,72 +98,43 @@ include(cmake/FetchQtGeneratorCMake.cmake)
 
 # ───── RESSOURCES ─────
 
-add_subdirectory(res)
+add_subdirectory(qml)
 
 # ───── QML SOURCES ─────
 
-file(GLOB_RECURSE QATERIAL_QML_SRCS qml/*.qml)
-
-foreach(SOURCE IN ITEMS ${QATERIAL_QML_SRCS})
-  get_filename_component(SOURCE_PATH "${SOURCE}" PATH)
-  file(RELATIVE_PATH SOURCE_PATH_REL ${CMAKE_CURRENT_SOURCE_DIR}/qml "${SOURCE_PATH}")
-  string(REPLACE "/" "\\" GROUP_PATH "${SOURCE_PATH_REL}")
-  source_group("Qml\\${GROUP_PATH}" FILES "${SOURCE}")
-endforeach()
-
-qt_generate_qmldir(QATERIAL_QMLDIR_FILE
- SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/qml/Qaterial
- MODULE "Qaterial")
-
-qt_generate_qrc(QATERIAL_QML_QRC
-  SOURCE_DIR ${CMAKE_CURRENT_SOURCE_DIR}/qml/Qaterial
-  DEST_DIR ${CMAKE_CURRENT_BINARY_DIR}/QaterialGenerated
-  NAME Qaterial.qrc
-  PREFIX "Qaterial"
-  GLOB_EXPRESSION "*.qml;qmldir"
-  ALWAYS_OVERWRITE
-  RECURSE)
-
-# Add qml to the qtquick compiler (should only be call once per target)
-qtquick_compiler_add_resources(QATERIAL_QML_RES ${QATERIAL_QML_QRC})
-
-# Dont Run Moc on cpp file generated from qml
-foreach(QATERIAL_CURRENT_QML_FILE ${QATERIAL_QML_RES})
-  set_property(SOURCE ${QATERIAL_CURRENT_QML_FILE} PROPERTY SKIP_AUTOMOC ON)
-  source_group("Qml\\QtQuickCompiler Files" FILES ${QATERIAL_CURRENT_QML_FILE})
-endforeach()
+set(QATERIAL_SOURCE_DIR src/Qaterial)
 
 set(QATERIAL_LAYOUT_SRCS
-  include/Qaterial/Layout.hpp
-  src/Layout.cpp
+  ${QATERIAL_SOURCE_DIR}/Layout.hpp
+  ${QATERIAL_SOURCE_DIR}/Layout.cpp
 )
 
 set(QATERIAL_IO_SRCS
-  include/Qaterial/TextFile.hpp
-  src/TextFile.cpp
-  include/Qaterial/Clipboard.hpp
-  src/Clipboard.cpp
-  include/Qaterial/FolderTreeModel.hpp
-  src/FolderTreeModel.cpp
+  ${QATERIAL_SOURCE_DIR}/TextFile.hpp
+  ${QATERIAL_SOURCE_DIR}/TextFile.cpp
+  ${QATERIAL_SOURCE_DIR}/Clipboard.hpp
+  ${QATERIAL_SOURCE_DIR}/Clipboard.cpp
+  ${QATERIAL_SOURCE_DIR}/FolderTreeModel.hpp
+  ${QATERIAL_SOURCE_DIR}/FolderTreeModel.cpp
 )
 
 set(QATERIAL_NAVIGATION_SRCS
-  include/Qaterial/StepperElement.hpp
-  src/StepperElement.cpp
-  include/Qaterial/TreeElement.hpp
-  src/TreeElement.cpp
+  ${QATERIAL_SOURCE_DIR}/StepperElement.hpp
+  ${QATERIAL_SOURCE_DIR}/StepperElement.cpp
+  ${QATERIAL_SOURCE_DIR}/TreeElement.hpp
+  ${QATERIAL_SOURCE_DIR}/TreeElement.cpp
 )
 
 set(QATERIAL_UTILS_SRCS
-  include/Qaterial/Qaterial.hpp
-  include/Qaterial/Export.hpp
-  include/Qaterial/Property.hpp
-  include/Qaterial/Version.hpp
-  src/Version.cpp
-  include/Qaterial/Utils.hpp
-  src/Utils.cpp
-  include/Qaterial/Logger.hpp
-  src/Logger.cpp
+  ${QATERIAL_SOURCE_DIR}/Qaterial.hpp
+  ${QATERIAL_SOURCE_DIR}/Export.hpp
+  ${QATERIAL_SOURCE_DIR}/Property.hpp
+  ${QATERIAL_SOURCE_DIR}/Version.hpp
+  ${QATERIAL_SOURCE_DIR}/Version.cpp
+  ${QATERIAL_SOURCE_DIR}/Utils.hpp
+  ${QATERIAL_SOURCE_DIR}/Utils.cpp
+  ${QATERIAL_SOURCE_DIR}/Logger.hpp
+  ${QATERIAL_SOURCE_DIR}/Logger.cpp
 )
 
 set(QATERIAL_SRCS
@@ -154,8 +142,6 @@ set(QATERIAL_SRCS
   ${QATERIAL_IO_SRCS}
   ${QATERIAL_NAVIGATION_SRCS}
   ${QATERIAL_UTILS_SRCS}
-  ${QATERIAL_QML_RES}
-  ${QATERIAL_QML_SRCS}
 )
 
 source_group("Layout" FILES ${QATERIAL_LAYOUT_SRCS})
@@ -174,14 +160,15 @@ else()
 endif()
 
 target_compile_features(${QATERIAL_TARGET} PUBLIC cxx_std_17)
-target_include_directories(${QATERIAL_TARGET} PUBLIC include)
+target_include_directories(${QATERIAL_TARGET} PUBLIC src)
 if(QATERIAL_ENABLE_PCH AND COMMAND target_precompile_headers)
-  target_precompile_headers(${QATERIAL_TARGET} PRIVATE include/Qaterial/Pch/Pch.hpp)
+  target_precompile_headers(${QATERIAL_TARGET} PRIVATE src/Qaterial/Pch/Pch.hpp)
 endif()
 
 target_link_libraries(${QATERIAL_TARGET} PRIVATE
-  ${QATERIAL_TARGET}Icons
-  ${QATERIAL_TARGET}Fonts
+  ${QATERIAL_TARGET}::Icons
+  ${QATERIAL_TARGET}::Fonts
+  ${QATERIAL_TARGET}::Components
 )
 
 target_link_libraries(${QATERIAL_TARGET} PUBLIC
@@ -196,7 +183,6 @@ target_link_libraries(${QATERIAL_TARGET} PUBLIC
   Qt::QuickControls2
 )
 
-set_target_properties(${QATERIAL_TARGET} PROPERTIES AUTORCC TRUE)
 set_target_properties(${QATERIAL_TARGET} PROPERTIES AUTOMOC TRUE)
 
 # Version
@@ -215,21 +201,9 @@ endif()
 # Correctly link to static qt
 get_target_property(QT_TARGET_TYPE Qt5::Core TYPE)
 if(${QT_TARGET_TYPE} STREQUAL "STATIC_LIBRARY")
-  include(cmake/FetchQtStaticCMake.cmake)
-  qt_generate_qml_plugin_import(${QATERIAL_TARGET}
-    QML_SRC qml
-    EXTRA_PLUGIN
-      QtQuickVirtualKeyboardPlugin
-      QtQuickVirtualKeyboardSettingsPlugin
-      QtQuickVirtualKeyboardStylesPlugin
-      QmlFolderListModelPlugin
-    VERBOSE
-  )
-  qt_generate_plugin_import(${QATERIAL_TARGET} VERBOSE)
-endif()
-
-if(TARGET Qt5::QmlWorkerScript)
-  target_link_libraries(${QATERIAL_TARGET} PUBLIC Qt5::QmlWorkerScript)
+  include(${PROJECT_SOURCE_DIR}/cmake/FetchQtStaticCMake.cmake)
+  qt_generate_qml_plugin_import(${QATERIAL_TARGET} QML_SRC ${CMAKE_CURRENT_SOURCE_DIR}/qml)
+  qt_generate_plugin_import(${QATERIAL_TARGET})
 endif()
 
 # ───── QATERIAL TOOLS ─────
