+++ cmake/Version.cmake
@@ -1,5 +1,5 @@
 set(QATERIAL_VERSION_MAJOR 1)
-set(QATERIAL_VERSION_MINOR 3)
+set(QATERIAL_VERSION_MINOR 4)
 set(QATERIAL_VERSION_PATCH 0)
 if(EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/.git")
   execute_process(
