--- include/Qaterial/Qaterial.hpp
@@ -1,9 +1,0 @@
-#ifndef __QATERIAL_HPP__
-#define __QATERIAL_HPP__
-
-#include <Qaterial/Export.hpp>
-#include <Qaterial/Version.hpp>
-#include <Qaterial/Utils.hpp>
-#include <Qaterial/Logger.hpp>
-
-#endif
