+++ qml/Qaterial/Icons/CMakeLists.txt
@@ -1,3 +1,25 @@
+# MIT License
+#
+# Copyright (c) 2020 Olivier Le Doeuff
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
 # No matter if QATERIAL_ENABLE_ICONS is ON or OFF, always generate something.
 # From Qaterial point of view nothing change
 
@@ -61,6 +83,7 @@ qt_generate_qrc(QATERIAL_QML_ICONS_ALIAS_QRC
   GLOB_EXPRESSION "*.qml;qmldir;")
 
 add_library(${QATERIAL_TARGET}Icons STATIC ${QATERIAL_QML_ICONS_QRC} ${QATERIAL_QML_ICONS_ALIAS_QRC})
+add_library(${QATERIAL_TARGET}::Icons ALIAS ${QATERIAL_TARGET}Icons)
 set_target_properties(${QATERIAL_TARGET}Icons PROPERTIES AUTORCC TRUE)
 
 target_link_libraries(${QATERIAL_TARGET}Icons PRIVATE
