--- res/CMakeLists.txt
@@ -1,2 +1,0 @@
-add_subdirectory(Fonts)
-add_subdirectory(Icons)
