--- src/_Template.cpp
@@ -1,16 +1,0 @@
-// ──── INCLUDE ────
-
-// Library Headers
-#include <Qaterial/.hpp>
-
-// Dependencies Headers
-
-// Qt Headers
-
-// Stl Headers
-
-// ──── DECLARATION ────
-
-using namespace qaterial;
-
-// ──── FUNCTIONS ────
